Consolidate errors into a single file & namespace errors under OpenCalais
