--- lib/calais_error/auth_error.rb
@@ -1,4 +1,0 @@
-module CalaisError
-  class AuthError < StandardError
-  end
-end
