--- lib/calais_error/bad_content.rb
@@ -1,4 +1,0 @@
-module CalaisError
-  class BadContent < StandardError
-  end
-end
