--- lib/calais_error/bad_media.rb
@@ -1,4 +1,0 @@
-module CalaisError
-  class BadMedia < StandardError
-  end
-end
