--- lib/calais_error/invalid_format.rb
@@ -1,4 +1,0 @@
-module CalaisError
-  class InvalidFormat < StandardError
-  end
-end
