--- lib/calais_error/large_entity.rb
@@ -1,5 +1,0 @@
-module CalaisError
-  class LargeEntity < StandardError
-
-  end
-end
