--- lib/calais_error/rate_limit_exceeded.rb
@@ -1,4 +1,0 @@
-module CalaisError
-  class RateLimitExceeded < StandardError
-  end
-end
