--- lib/calais_error/resource_not_found.rb
@@ -1,4 +1,0 @@
-module CalaisError
-  class ResourceNotFound < StandardError
-  end
-end
