--- lib/calais_error/server_error.rb
@@ -1,4 +1,0 @@
-module CalaisError
-  class ServerError < StandardError
-  end
-end
