--- lib/calais_error/service_unavailable.rb
@@ -1,4 +1,0 @@
-module CalaisError
-  class ServiceUnavailable < StandardError
-  end
-end
