+++ lib/open_calais/client.rb
@@ -3,7 +3,7 @@
 require 'open_calais/configuration'
 require 'open_calais/connection'
 require 'open_calais/response'
-require 'errors/raise_error'
+require 'open_calais/errors'
 
 module OpenCalais
   class Client
